requirements.txt
asgiref
certifi
charset-normalizer
APScheduler
django_apscheduler
django-request
dj_database_url
django-request
Django==4.0.5
gunicorn
idna
matplotlib
numpy
pip
psycopg2
python-dateutil
python-dotenv
requests
setuptools
six
sqlparse
tweepy
tzdata
urllib3
wheel
whitenoise
django-widget-tweaks
